Seperate empty redox response body
CastException was thrown when EmptyResponse was tried to cast to MetaResponse in
the case of empty redox response body. This case was seperated and an error
message is returned using Left

Including the empty redox body case to the successful cases just because it has
a success status does not make much sense. It is also not a complete failure, it
is just an unwanted behaviour, so seperating its case was decided to be a better
solution.

Consume should decide what to do with such cases. EmptyResponse extending
MetaResponse did not make much sense because Meta field of MetaResponse needed
to be invalidated.