Separate Clinical Summary and events from the rest
Initially, all redox events were implemented using hierarchy. @Common.scala exists for fields, but event types also shared patient, demographics, headers etc. which made it difficult to change fields for specific events when redox changes them on their side. This situation happened when Redox made it so that email address field for Clinical Summary used array of JSObjects and other data types used array of strings.

Separated Clinical Summary and its event types completely from the rest. They still share common fields like Meta, Identifier, Address etc. but fields like demographics or patient is separated so that it is flexible to change just Clinical Summary and not others. This is also meaningful change because inheritance is not suited for request/response models. This was a legacy approach that will be changed over time as models need to be separated from others.